+++ internal/provider/fwprovider/provider.go
@@ -242,7 +242,7 @@ func (p *fwprovider) Resources(ctx context.Context) []func() resource.Resource {
 		spiffesec.NewSpiffeSecretBackendConfigResource,
 		spiffesec.NewSpiffeSecretBackendRoleResource,
 		radius.NewRadiusAuthBackendResource,
-
+		radius.NewRadiusAuthBackendUserResource,
 		sys.NewPasswordPolicyResource,
 		sys.NewRotationPolicyResource,
 		sysconfig.NewQuotaConfigResource,